refactor of getUsers. get token and get users requests in different modules. Rename Getusers component
